Improve type declarations and add TSDoc
