+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.3.1] - 2020-10-02
+### Added
+- Improved type exports (`/types`)
+- Added TSDoc comments
+
 ## [0.3.0] - 2020-10-01
 ### Removed
 - Removed the `FormCreateMixin` (replaced with typed `createForm` function)
