+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@kendallroth/vue-simple-forms",
-  "version": "0.3.0",
+  "version": "0.3.1",
   "description": "Simple Vue form state management library",
   "main": "./lib/index.js",
   "types": "./lib/index.d.ts",
@@ -83,7 +83,7 @@
     }
   },
   "lint-staged": {
-    "{src,test}/**/*.js": [
+    "{src,test}/**/*.ts": [
       "eslint --fix",
       "prettier --write"
     ]
